[dice] move TPM cert gen functions to separate lib
This adds a TPM lib for generating TPM certificates, moving them out of
the DICE lib, as these certificates stem from a separate set of keys.
