+++ sw/device/silicon_creator/lib/BUILD
@@ -543,36 +543,6 @@ cc_library(
 )
 
 cc_library(
-    name = "dice",
-    srcs = ["dice.c"],
-    hdrs = ["dice.h"],
-    target_compatible_with = [OPENTITAN_CPU],
-    deps = [
-        "//hw/ip/otp_ctrl/data:otp_ctrl_c_regs",
-        "//sw/device/lib/base:memory",
-        "//sw/device/lib/base:status",
-        "//sw/device/lib/testing/json:provisioning_data",
-        "//sw/device/lib/testing/test_framework:check",
-        "//sw/device/silicon_creator/lib:attestation",
-        "//sw/device/silicon_creator/lib:attestation_key_diversifiers",
-        "//sw/device/silicon_creator/lib:otbn_boot_services",
-        "//sw/device/silicon_creator/lib/base:util",
-        "//sw/device/silicon_creator/lib/cert",
-        "//sw/device/silicon_creator/lib/cert:cdi_0_template_library",
-        "//sw/device/silicon_creator/lib/cert:cdi_1_template_library",
-        "//sw/device/silicon_creator/lib/cert:tpm_cek_template_library",
-        "//sw/device/silicon_creator/lib/cert:tpm_cik_template_library",
-        "//sw/device/silicon_creator/lib/cert:tpm_ek_template_library",
-        "//sw/device/silicon_creator/lib/cert:uds_template_library",
-        "//sw/device/silicon_creator/lib/drivers:hmac",
-        "//sw/device/silicon_creator/lib/drivers:keymgr",
-        "//sw/device/silicon_creator/lib/drivers:lifecycle",
-        "//sw/device/silicon_creator/lib/drivers:otp",
-        "//sw/device/silicon_creator/lib/sigverify:ecdsa_p256_key",
-    ],
-)
-
-cc_library(
     name = "stack_utilization",
     srcs = ["stack_utilization.c"],
     hdrs = [
