+++ sw/device/silicon_creator/lib/cert/dice.h
@@ -2,12 +2,11 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef OPENTITAN_SW_DEVICE_SILICON_CREATOR_LIB_DICE_H_
-#define OPENTITAN_SW_DEVICE_SILICON_CREATOR_LIB_DICE_H_
+#ifndef OPENTITAN_SW_DEVICE_SILICON_CREATOR_LIB_CERT_DICE_H_
+#define OPENTITAN_SW_DEVICE_SILICON_CREATOR_LIB_CERT_DICE_H_
 
 #include <stdint.h>
 
-#include "sw/device/lib/testing/json/provisioning_data.h"
 #include "sw/device/silicon_creator/lib/attestation.h"
 #include "sw/device/silicon_creator/lib/cert/cert.h"
 #include "sw/device/silicon_creator/lib/drivers/hmac.h"
@@ -21,11 +20,6 @@ enum {
    */
   kDiceMeasurementSizeInBits = 256,
   kDiceMeasurementSizeInBytes = kDiceMeasurementSizeInBits / 8,
-
-  /**
-   * DICE key ID size (for use in DICE certificates).
-   */
-  kDiceCertKeyIdSizeInBytes = 20,
 };
 
 /**
@@ -34,9 +28,6 @@ enum {
 extern const sc_keymgr_ecc_key_t kDiceKeyUds;
 extern const sc_keymgr_ecc_key_t kDiceKeyCdi0;
 extern const sc_keymgr_ecc_key_t kDiceKeyCdi1;
-extern const sc_keymgr_ecc_key_t kDiceKeyTpmEk;
-extern const sc_keymgr_ecc_key_t kDiceKeyTpmCek;
-extern const sc_keymgr_ecc_key_t kDiceKeyTpmCik;
 
 /**
  * Generates the UDS attestation keypair and (unendorsed) X.509 TBS certificate.
@@ -96,54 +87,4 @@ rom_error_t dice_cdi_1_cert_build(hmac_digest_t *owner_measurement,
                                   ecdsa_p256_public_key_t *cdi_1_pubkey,
                                   uint8_t *cert, size_t *cert_size);
 
-/**
- * Generates an X.509 TBS section of a TPM EK certificate.
- *
- * @param key_ids Pointer to the (current and endorsement) public key IDs.
- * @param tpm_ek_pubkey Pointer to the TPM EK public key in big endian.
- * @param[out] tbs_cert Buffer to hold the generated TBS section.
- * @param[in,out] tbs_cert_size Size of the generated TBS section (input value
- *                              is the size of the allocated cert_buf, output
- *                              value final computed size of the certificate).
- * @return The result of the operation.
- */
-OT_WARN_UNUSED_RESULT
-rom_error_t dice_tpm_ek_tbs_cert_build(cert_key_id_pair_t *key_ids,
-                                       ecdsa_p256_public_key_t *tpm_ek_pubkey,
-                                       uint8_t *tbs_cert,
-                                       size_t *tbs_cert_size);
-
-/**
- * Generates an X.509 TBS section of a TPM CEK certificate.
- *
- * @param key_ids Pointer to the (current and endorsement) public key IDs.
- * @param tpm_cek_pubkey Pointer to the TPM CEK public key in big endian.
- * @param[out] tbs_cert Buffer to hold the generated UDS section.
- * @param[in,out] tbs_cert_size Size of the generated TBS section (input value
- *                              is the size of the allocated cert_buf, output
- *                              value final computed size of the certificate).
- * @return The result of the operation.
- */
-OT_WARN_UNUSED_RESULT
-rom_error_t dice_tpm_cek_tbs_cert_build(cert_key_id_pair_t *key_ids,
-                                        ecdsa_p256_public_key_t *tpm_cek_pubkey,
-                                        uint8_t *tbs_cert,
-                                        size_t *tbs_cert_size);
-
-/**
- * Generates an X.509 TBS section of a TPM CIK certificate.
- *
- * @param key_ids Pointer to the (current and endorsement) public key IDs.
- * @param tpm_cik_pubkey Pointer to the TPM CIK public key in big endian.
- * @param[out] tbs_cert Buffer to hold the generated UDS section.
- * @param[in,out] tbs_cert_size Size of the generated TBS section (input value
- *                              is the size of the allocated cert_buf, output
- *                              value final computed size of the certificate).
- * @return The result of the operation.
- */
-OT_WARN_UNUSED_RESULT
-rom_error_t dice_tpm_cik_tbs_cert_build(cert_key_id_pair_t *key_ids,
-                                        ecdsa_p256_public_key_t *tpm_cik_pubkey,
-                                        uint8_t *tbs_cert,
-                                        size_t *tbs_cert_size);
-#endif  // OPENTITAN_SW_DEVICE_SILICON_CREATOR_LIB_DICE_H_
+#endif  // OPENTITAN_SW_DEVICE_SILICON_CREATOR_LIB_CERT_DICE_H_
