+++ sw/device/silicon_creator/manuf/base/BUILD
@@ -230,12 +230,12 @@ cc_library(
         "//sw/device/lib/testing/test_framework:ujson_ottf",
         "//sw/device/silicon_creator/lib:attestation",
         "//sw/device/silicon_creator/lib:attestation_key_diversifiers",
-        "//sw/device/silicon_creator/lib:dice",
         "//sw/device/silicon_creator/lib:otbn_boot_services",
         "//sw/device/silicon_creator/lib/base:util",
         "//sw/device/silicon_creator/lib/cert",
         "//sw/device/silicon_creator/lib/cert:cdi_0_template_library",
         "//sw/device/silicon_creator/lib/cert:cdi_1_template_library",
+        "//sw/device/silicon_creator/lib/cert:dice",
         "//sw/device/silicon_creator/lib/cert:tpm_cek_template_library",
         "//sw/device/silicon_creator/lib/cert:tpm_cik_template_library",
         "//sw/device/silicon_creator/lib/cert:tpm_ek_template_library",
@@ -258,8 +258,9 @@ cc_library(
         "//sw/device/lib/runtime:log",
         "//sw/device/lib/testing/test_framework:status",
         "//sw/device/lib/testing/test_framework:ujson_ottf",
-        "//sw/device/silicon_creator/lib:dice",
+        "//sw/device/silicon_creator/lib:attestation",
         "//sw/device/silicon_creator/lib/cert",
+        "//sw/device/silicon_creator/lib/cert:tpm",
         "//sw/device/silicon_creator/lib/cert:tpm_cek_template_library",
         "//sw/device/silicon_creator/lib/cert:tpm_cik_template_library",
         "//sw/device/silicon_creator/lib/cert:tpm_ek_template_library",
